Updates to Implement XInput and support for MinGW-w32
This version allows usage of XInput to support XBox controllers
specifically, allowing the left and right triggers to be parsed
separately. Also added a button enumeration that includes the left and
right triggers so that buttons can be tested for presses with the
enumeration values.
This version must be built with MinGW-w64 and MinGW-w32, since only that
version of gcc allows linking against xinput (even though the header
file is in this project and comes from the Windows Platform SDK).